docker: add build-push script, include withdrawal-backfill in image, fix uint64 type
- build-push-docker.sh: multi-platform build and push (bosagora/agora-scan)
- Makefile: add withdrawal-backfill target to 'make all'
- Dockerfile: chmod withdrawal-backfill binary
- cmd/withdrawal-backfill: cast flag.Uint to uint64 for runBackfill

+++ cmd/withdrawal-backfill/main.go
@@ -54,7 +54,7 @@ func main() {
 
 	for {
 		t0 := time.Now()
-		runBackfill(slotsPerEpoch, *lookback)
+		runBackfill(slotsPerEpoch, uint64(*lookback))
 		elapsed := time.Since(t0)
 		logrus.Infof("backfill cycle finished in %v, sleeping %v", elapsed, *interval)
 		time.Sleep(*interval)
